+++ src/pages/Detail.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useState } from "react";
 import { useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import { octokitDetailAPI } from "../api/client";
+import { octokitDetailAPI } from "../api/issue";
 import ReactMarkdown from "react-markdown";
 import {
   IssueContainer,
